cleanup: moved welcome component outside of ui folder
